Add Kafka, Hive and updated others

+++ spark.sh
@@ -1,6 +1,10 @@
 # Navigate to home directory and clear up folders and directories
 cd
-sudo rm spark-*
+FOLDER=`ls -d */ | grep spark-`
+if [ -d "${FOLDER}" ]; then
+    echo "Deleting previous spark installations"
+    sudo rm spark-*
+fi
 sudo rm -rf spark-*
 
 
@@ -9,16 +13,21 @@ sudo apt install scala git -y
 
 
 # Downloading and setting up Spark
-wget https://downloads.apache.org/spark/spark-3.1.2/spark-3.1.2-bin-hadoop3.2.tgz
-tar xzf spark-3.1.2-bin-hadoop3.2.tgz
-sudo mv spark-3.1.2-bin-hadoop3.2 /opt/spark
-
+wget https://dlcdn.apache.org/spark/spark-3.3.0/spark-3.3.0-bin-hadoop3.tgz
+tar xzf spark-3.3.0-bin-hadoop3.tgz
+if [ -d "/opt/spark" ]; then
+    echo "Spark directory exists, deleting it."
+    sudo rm -rf /opt/spark
+fi
+sudo mv spark-3.3.0-bin-hadoop3 /opt/spark
+sudo rm -rf spark-*
 
 # Configuring Spark
-sudo echo "export SPARK_HOME=/opt/spark" >> ~/.profile
-sudo echo "export PATH=$PATH:$SPARK_HOME/bin:$SPARK_HOME/sbin" >> ~/.profile
-sudo echo "export PYSPARK_PYTHON=/usr/bin/python3" >> ~/.profile
-sudo source ~/.profile
+echo "export SPARK_HOME=/opt/spark" >> ~/.bashrc
+source ~/.bashrc
+echo "export PATH=$PATH:$SPARK_HOME/bin:$SPARK_HOME/sbin" >> ~/.bashrc
+echo "export PYSPARK_PYTHON=/usr/bin/python3" >> ~/.bashrc
+source ~/.bashrc
 
 # Starting Spark
 cd /opt/spark/sbin
@@ -29,5 +38,14 @@ sudo chmod +x pyspark
 sudo chmod +x spark-shell
 sudo chmod +x spark-submit
 
+cd ../sbin
 
 ./start-all.sh
+
+cd ..
+
+./bin/run-example SparkPi 10
+
+cd sbin
+
+./stop-all.sh
